app-etl/requirements.txt
numpy==1.23.0
pandas==1.5.3
pyarrow==12.0.1
psycopg2-binary==2.9.6
minio==7.1.6
